RaspberryPythonRaceController_container/requirements.txt
asyncio-mqtt==0.16.2
paho-mqtt==1.6.1
pyserial==3.5
boto3==1.34.34
botocore==1.34.34
cryptography==42.0.2
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-mock==3.12.0